# core/trie: persist TrieJournal to journal file instead of kv database