--- a/cmd/geth/snapshot.go
+++ b/cmd/geth/snapshot.go
@@ -24,7 +24,7 @@ import (
 	"os"
 	"time"
 
-	cli "github.com/urfave/cli/v2"
+	"github.com/urfave/cli/v2"
 
 	"github.com/ethereum/go-ethereum/cmd/utils"
 	"github.com/ethereum/go-ethereum/common"