--- a/eth/backend.go
+++ b/eth/backend.go
@@ -207,10 +207,8 @@ func New(stack *node.Node, config *ethconfig.Config) (*Ethereum, error) {
 		journalFilePath string
 		path            string
 	)
-	if config.JournalFileEnabled {
-		path = ChainData
-		journalFilePath = stack.ResolvePath(path) + "/" + JournalFileName
-	}
+	path = ChainData
+	journalFilePath = stack.ResolvePath(path) + "/" + JournalFileName
 	var (
 		vmConfig = vm.Config{
 			EnablePreimageRecording:   config.EnablePreimageRecording,
@@ -233,6 +231,7 @@ func New(stack *node.Node, config *ethconfig.Config) (*Ethereum, error) {
 			EnableProofKeeper:    config.EnableProofKeeper,
 			KeepProofBlockSpan:   config.KeepProofBlockSpan,
 			JournalFilePath:      journalFilePath,
+			JournalFile:          config.JournalFileEnabled,
 		}
 	)
 	// Override the chain config with provided settings.