--- a/eth/ethconfig/config.go
+++ b/eth/ethconfig/config.go
@@ -134,6 +134,7 @@ type Config struct {
 	ProposeBlockInterval uint64                `toml:",omitempty"` // Keep the same with op-proposer propose block interval
 	EnableProofKeeper    bool                  `toml:",omitempty"` // Whether to enable proof keeper
 	KeepProofBlockSpan   uint64                `toml:",omitempty"` // Span block of keep proof
+	JournalFileEnabled   bool                  // Whether the TrieJournal is stored using journal file
 
 	// RequiredBlocks is a set of block number -> hash mappings which must be in the
 	// canonical chain of all remote peers. Setting the option makes geth verify the