--- a/trie/triedb/pathdb/database.go
+++ b/trie/triedb/pathdb/database.go
@@ -20,6 +20,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"os"
 	"sync"
 	"time"
 
@@ -101,6 +102,7 @@ type Config struct {
 	ReadOnly             bool           // Flag whether the database is opened in read only mode.
 	ProposeBlockInterval uint64         // Propose block to L1 block interval.
 	NotifyKeep           NotifyKeepFunc // NotifyKeep is used to keep the proof which maybe queried by op-proposer.
+	JournalFilePath      string
 }
 
 // sanitize checks the provided user configurations and changes anything that's
@@ -320,7 +322,7 @@ func (db *Database) Enable(root common.Hash) error {
 	// Drop the stale state journal in persistent database and
 	// reset the persistent state id back to zero.
 	batch := db.diskdb.NewBatch()
-	rawdb.DeleteTrieJournal(batch)
+	db.DeleteTrieJournal(batch)
 	rawdb.WritePersistentStateID(batch, 0)
 	if err := batch.Write(); err != nil {
 		return err
@@ -386,7 +388,7 @@ func (db *Database) Recover(root common.Hash, loader triestate.TrieLoader) error
 		// disk layer won't be accessible from outside.
 		db.tree.reset(dl)
 	}
-	rawdb.DeleteTrieJournal(db.diskdb)
+	db.DeleteTrieJournal(db.diskdb)
 	_, err := truncateFromHead(db.diskdb, db.freezer, dl.stateID())
 	if err != nil {
 		return err
@@ -506,6 +508,23 @@ func (db *Database) modifyAllowed() error {
 	}
 	if db.waitSync {
 		return errDatabaseWaitSync
+	}
+	return nil
+}
+
+func (db *Database) DeleteTrieJournal(writer ethdb.KeyValueWriter) error {
+	filePath := db.config.JournalFilePath
+	if len(filePath) == 0 {
+		rawdb.DeleteTrieJournal(writer)
+	} else {
+		_, err := os.Stat(filePath)
+		if os.IsNotExist(err) {
+			return err
+		}
+		errRemove := os.Remove(filePath)
+		if errRemove != nil {
+			log.Crit("Failed to remove tries journal", "journal path", filePath, "err", err)
+		}
 	}
 	return nil
 }