--- a/trie/triedb/pathdb/nodebufferlist.go
+++ b/trie/triedb/pathdb/nodebufferlist.go
@@ -727,7 +727,7 @@ func (w *proposedBlockReader) parentLayer() layer { return nil }
 func (w *proposedBlockReader) update(root common.Hash, id uint64, block uint64, nodes map[common.Hash]map[string]*trienode.Node, states *triestate.Set) *diffLayer {
 	return nil
 }
-func (w *proposedBlockReader) journal(io.Writer, bool) error { return nil }
+func (w *proposedBlockReader) journal(io.Writer, JournalType) error { return nil }
 
 // multiDifflayer compresses several difflayers in one map. As an element of nodebufferlist
 // it is the smallest unit for storing trie nodes.