### Description

This PR provides another way to persist TrieJournal by write it into a log file instead of a kv database.

Use `--journalFile` flag to enable/disable the journal file feature.

### Rationale

Persisting TrieJournal to KV database will cause problems below

- Additional WAL and compaction overhead in pebble/leveldb
- Big kv (>= 2GB) not supported by pebbledb. ([Size out of bounds when write kv >= 2GB](https://github.com/cockroachdb/pebble/issues/3333))

### Changes:

**Data Format**

Journal Version
Disk Root
Disk Layer
- Root
- State ID
- Buffer Nodes
- `Layer Checksum`

Diff Layer

- Root
- Block
- Trie Nodes
- Journal Accounts
- `Layer Checksum`

Journal Path: geth/chaindata/ancient/state.journal
Checksum: SHA256

####  Performance test

Scenario: Using TestJournal, each layer has 4K and 8K account operations.
|            | WriteJournal | LoadJournal | Dirty State Size | 
|------------|--------------|-------------|------------------| 
| **New Journal WAL** | 6.149s | 12.496s | 1.19GiB | 
| Old KVDB        | 6.341s | 12.340s | 1.21GiB | 
| **New Journal WAL** | 11.724s | 20.525s | 1.85GiB | 
| Old KVDB        | 11.107s | 19.415s | 1.85GiB |  

Conclusion: The use of Journal WAL and storing trieJournalKey directly into kvdb results in nearly identical WAL write times during shutdown and WAL read/parsing times during startup.